# Gcd deserves to be simpler and faster.