--- a/lib/std/math/gcd.zig
+++ b/lib/std/math/gcd.zig
@@ -9,42 +9,26 @@ pub fn gcd(a: anytype, b: anytype) @TypeOf(a, b) {
         comptime_int => std.math.IntFittingRange(@min(a, b), @max(a, b)),
         else => |T| T,
     };
+
     if (@typeInfo(N) != .int or @typeInfo(N).int.signedness != .unsigned) {
         @compileError("`a` and `b` must be usigned integers");
     }
 
-    // using an optimised form of Stein's algorithm:
-    // https://en.wikipedia.org/wiki/Binary_GCD_algorithm
-    std.debug.assert(a != 0 or b != 0);
-
-    if (a == 0) return b;
-    if (b == 0) return a;
-
+    // Euclidean algorithm.
     var x: N = a;
     var y: N = b;
 
-    const xz = @ctz(x);
-    const yz = @ctz(y);
-    const shift = @min(xz, yz);
-    x >>= @intCast(xz);
-    y >>= @intCast(yz);
-
-    var diff = y -% x;
-    while (diff != 0) : (diff = y -% x) {
-        // ctz is invariant under negation, we
-        // put it here to ease data dependencies,
-        // makes the CPU happy.
-        const zeros = @ctz(diff);
-        if (x > y) diff = -%diff;
-        y = @min(x, y);
-        x = diff >> @intCast(zeros);
+    while (x != 0) {
+        const r = y % x;
+        y = x;
+        x = r;
     }
-    return y << @intCast(shift);
+
+    return y;
 }
 
 test gcd {
     const expectEqual = std.testing.expectEqual;
-
     try expectEqual(gcd(0, 5), 5);
     try expectEqual(gcd(5, 0), 5);
     try expectEqual(gcd(8, 12), 4);