The current implementation uses Stein's algorithm.

This algorithm is outdated, because modern cpu's are smart at optimizing the modulo operation, i.e. the "tricks" that are applied in the algorithm explicitly (i.e. shifting, comparing count of trailing zeros), are usually done by the cpu already and should be avoided if possible.

My local benchmarks indicate that simply using the Euclidean algorithm is almost 2x faster than using Stein's algorithm for random `u64` inputs.

Warmup iterations: 500k.
Timed iterations: 5kk.

Some random run 1:
```
Time needed (.none): 3262ms   // cost of generating random u64s.
Time needed (.euclid): 4517ms // +~1.3sec
Time needed (.steins): 6206ms // +~2.9sec
```

Some random run 2:
```
Time needed (.none): 3171ms
Time needed (.euclid): 3935ms // +~0.76
Time needed (.steins): 4832ms // +~1.66
```

On a side note: both algorithms require `~1.44 * n` steps of iterations in the worst case (`n` is bits of smallest number).
It is possible to reduce this to `~0.79 *n` steps in the worst case. This algorithm outperforms the Euclidean algorithm by almost 2x speed for worst case inputs, but for random inputs it is a bit worse because it has one additional branch. If interested, I can also add this algorithm -- called "Ring" -- in the gcd file.