@Fri3dNstuff 2025-09-27
Can you share your testing methodology? `std.math.gcd` used Euclid's algorithm in the past, and was changed to use Stein's due to better performance, which was benchmarked. https://github.com/ziglang/zig/pull/21077