@two-horned 2025-09-27
Sure. It is nothing too special, but it gets the job done.

```zig
const std = @import("std");
const gcd_ring = @import("gcd.zig").gcd;
const gcd_old = @import("gcd_old.zig").gcd;
const gcd_euclid = @import("gcd_euclid.zig").gcd;

fn none(a: anytype, b: anytype) @TypeOf(a, b) {
    const N = switch (@TypeOf(a, b)) {
        // convert comptime_int to some sized int type for @ctz
        comptime_int => std.math.IntFittingRange(@min(a, b), @max(a, b)),
        else => |T| T,
    };

    if (@typeInfo(N) != .int or @typeInfo(N).int.signedness != .unsigned) {
        @compileError("`a` and `b` must be usigned integers");
    }
    return @max(a, b);
}

const warmup_iterations = 500_000;
const totest_iterations = 5_000_000;

pub fn main() !void {
    var prng = std.Random.DefaultPrng.init(blk: {
        var seed: u64 = undefined;
        try std.posix.getrandom(std.mem.asBytes(&seed));
        break :blk seed;
    });
    const rand = prng.random();
    var timer = try std.time.Timer.start();

    const gcd_ids = enum(usize) { none = 0, ring = 1, steins = 2, euclid = 3 };
    const gcd_fns = .{ none, gcd_ring, gcd_old, gcd_euclid };

    inline for (0..gcd_fns.len) |i| {
        const gcd = gcd_fns[i];
        for (0..warmup_iterations) |_| {
            const a = rand.int(u64);
            const b = rand.int(u64);
            _ = gcd(a, b);
        }

        timer.reset();
        for (0..totest_iterations) |_| {
            const a = rand.int(u64);
            const b = rand.int(u64);
            std.debug.assert(gcd(a, b) == gcd(b, a));
        }
        const now = timer.read() / std.time.ns_per_ms;
        const id: gcd_ids = @enumFromInt(i);
        std.debug.print("Time needed ({}): {}ms\n", .{ id, now });
    }
    inline for (0..gcd_fns.len) |j| {
        const i = gcd_fns.len - j - 1;
        const gcd = gcd_fns[i];
        for (0..warmup_iterations) |_| {
            const a = rand.int(u64);
            const b = rand.int(u64);
            std.debug.assert(gcd(a, b) == gcd(b, a));
        }

        timer.reset();
        for (0..totest_iterations) |_| {
            const a = rand.int(u64);
            const b = rand.int(u64);
            _ = gcd(a, b);
        }
        const now = timer.read() / std.time.ns_per_ms;
        const id: gcd_ids = @enumFromInt(i);
        std.debug.print("Time needed ({}): {}ms\n", .{ id, now });
    }
}
```