@Fri3dNstuff 2025-09-27
@two-horned, can you also share the definitions of the different implementations, please?
```zig
const gcd_ring = @import("gcd.zig").gcd;
const gcd_old = @import("gcd_old.zig").gcd;
const gcd_euclid = @import("gcd_euclid.zig").gcd;
```