@two-horned 2025-09-27
`gcd_euclid.zig` is the [updated gcd file](https://github.com/ziglang/zig/blob/12793521fe76f76d82ced0d47ffe151a25dac758/lib/std/math/gcd.zig) in my commit. The "ring" is not relevant for now, but I can share it if you want.

What I noticed is, debug.assert(a != 0 or b != 0) is missing, but even after inserting this assert, my results are comparable to the ones above. Also, I think the definition of gcd(0,0) = 0 is OK and doesn't need to panic.