@two-horned 2025-09-27
What I ran was `zig build-exe -Doptimize=ReleaseFast main.zig && ./main`,
but `zig run -DReleaseFast main.zig` gives the same results.

I updated the above test code to exclude ring.