@Fri3dNstuff 2025-09-27
The flag to make a release-fast build is `-OReleaseFast` (notice - capital 'O') - a `-D` flag, as per my understanding, is to define C macros. From `zig build-exe --help`:
```
Per-Module Compile Options:
  -target [name]            <arch><sub>-<os>-<abi> see the targets command
  -O [mode]                 Choose what to optimize for
    Debug                   (default) Optimizations off, safety on
    ReleaseFast             Optimize for performance, safety off
    ReleaseSafe             Optimize for performance, safety on
    ReleaseSmall            Optimize for small binary, safety off
    
    ✂️ sinpped lines ✂️
    
  -D[macro]=[value]         Define C [macro] to [value] (1 if [value] omitted)
```
Perhaps you are confusing it with the flag provided to a `zig build` command?