@two-horned 2025-09-27
Interestingly, when I run on version 0.16.0, the test results is actually faster for all cases.
But for me the Euclidean algorithm still performs better.

```
❯ zig run -DReleaseFast main_share.zig
Time needed (.none): 3199ms
Time needed (.steins): 6376ms
Time needed (.euclid): 5170ms
Time needed (.euclid): 4193ms
Time needed (.steins): 4815ms
Time needed (.none): 3182ms
❯ ~/.local/src/zig-x86_64-linux-0.16.0-dev.393+dd4be26f5/zig run -DReleaseFast main_share.zig
Time needed (.none): 2054ms
Time needed (.steins): 5153ms
Time needed (.euclid): 4017ms
Time needed (.euclid): 3037ms
Time needed (.steins): 3652ms
Time needed (.none): 1997ms
```