@Fri3dNstuff 2025-09-27
Please try to run the benchmark in the release-fast build mode, using `-OReleaseFast` (after the dash comes a capital 'O', 15th letter of the Basic Latin Alphabet - not 'D', 4th letter of the alphabet).