@two-horned 2025-09-27
I see, the `-Doptimize=ReleaseSmall` is something that only works for projects? Very confusing.

My benchmarks align with yours after using `-OReleaseFast`.